Polynomial Regression
Let’s assume the above dateset table is what the HR team of a company uses to determine what salary to offer to a new employee. For example that an employee has applied for the role of a Regional Manager and has already worked as a Regional Manager for 2 years. So based on the table — he falls between level 6 and level 7 — Lets say he falls at level 6.5
Build a model to predict what salary we should offer this new employee.